backend/requirements.txt
flask
flask_cors
python-dotenv
streamlit
gradio
langchain
langchain-community
google-generativeai
langchain-google-genai
pypdf2
beautifulsoup4
unstructured
nltk
langchain-groq
faiss-cpu
pandas
openpyxl
python-docx
reportlab
fpdf